# HostFunc: remove reflect for host function call