This PR remove reflect for wasm host function call, the benchmark is:

**Before：**
``` 
goos: darwin
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkHostFunctionCall
BenchmarkHostFunctionCall/call_go_host
BenchmarkHostFunctionCall/call_go_host-8                  939739              1262 ns/op
BenchmarkHostFunctionCall/call_wasm_host
BenchmarkHostFunctionCall/call_wasm_host-8              22812769                48.90 ns/op
PASS
ok      github.com/tetratelabs/wazero/internal/integration_test/bench   2.658s
```

**After:**

```
goos: darwin
goarch: amd64
pkg: github.com/tetratelabs/wazero/examples/import-go
BenchmarkHostFunctionCall
BenchmarkHostFunctionCall/call_go_host
BenchmarkHostFunctionCall/call_go_host-8                 9719632               125.6 ns/op
BenchmarkHostFunctionCall/call_wasm_host
BenchmarkHostFunctionCall/call_wasm_host-8              24260326                48.20 ns/op
PASS
ok      github.com/tetratelabs/wazero/examples/import-go        2.988s
```