@codefromthecrypt 2022-10-20
Thanks I see what you are up to. Are you interested in proceeding? If so, I can wait until it is all compiling and then perform review and/or help polish.