@codefromthecrypt 2022-10-20
ps my advice would be to implement HostFuncSignature with reflection by default. Ex if the input function is not a HostFuncSignature, then make a HostFuncSignature which uses the existing reflective logic.