@hafeidejiangyou 2022-10-21
@codefromthecrypt  You mean support two way to write host function at the same time? 

>  if the input function is not a HostFuncSignature, then make a HostFuncSignature which uses the existing reflective logic.

make a HostFuncSignature for call the non HostFuncSignature function is ok, but it will use reflect to do the final call.
In this way, the benefit is user can write host function with more go type not just uint64, is my understanding correct?