@codefromthecrypt 2022-10-21
@hafeidejiangyou let's call the type `api.HostFunc`

ex.
```go
package api

type HostFunc func(ctx context.Context, mod api.Module, in ...uint64) ([]uint64, error)
```

We then move the field `wasm.Code.GoFunc` to `wasm.HostFunc` and add `wasm.Code.HostFunc` as the above type.
The main idea is to make sure all code paths don't use reflection.

We add `wasm.HostFunc.GoFunc` for transition purposes explained below

We add a `HostFuncBuilder` which collects this and the metadata it needs.
```go
package wazero

type HostFuncBuilder interface {
	WithParams(api.ValueType...) HostFuncBuilder

	WithResults(api.ValueType...) HostFuncBuilder

	Build() interface{} // or any, but not the internal type.
}

type hostFuncBuilder struct {
	fn api.HostFunc
	params, values []api.ValueType
}

func NewHostFuncExporter(fn HostFunc) {
--snip--

func (b *hostFuncBuilder) Build() {
	return &wasm.HostFunc{Code: wasm.Code{HostFunc: b.fn ...} // note the new field HostFunc
}
```

HostModuleBuilder is left alone.. Let's take example of two calls to  `HostModuleBuilder.ExportFunction`
```
		ExportFunction("add_1_reflect", func(in uint32) uint32 {
				return in+1
			}).
		ExportFunction("add_1_hostfunc", wazero.NewHostFuncBuilder(
			func(_context.Context,_ api.Module, params ...uint64) ([]uint64, error) {
				return []uint64{params[0]+1}, nil
			}).
                        WithParams(api.ValueTypeI32).
                        WithResults(api.ValueTypeI32).
                        Build()).
```

So HostModuleBuilder.Build() ends up with two entries in the map: 

* "add_1_reflect": the anonymous func instance
* "add_1_hostfunc": an instance of wasm.HostFunc which has a new field of `api.HostFunc` set.
  * Note: we already have internal code that add  `wasm.HostFunc` to the builder. ex assemblyscript!

So all the same logic is used... this map is passed to `wasm.NewHostModule` same as now, but the logic in `addFuncs` changes.

Where it has this.. add something to adapt as necessary
```
if hf, ok := v.(*HostFunc); ok {
	if gf := hf.GoFunc; gf != nil {
		hf.Code.HostFunc = parseGoFunc(gf)
		...
		hf.GoFunc = nil // done
```

The main thing is to change the logic here to make sure that if HostFunc.GoFunc is present it is adapted to an api.GoFunc (using reflection), but making sure the engine is agnostic to this.