@hafeidejiangyou 2022-10-24
@codefromthecrypt  I think you have wrote almost all the code, to reduce communication costs and review costs, I think you can modify it directly, if you think it is needed.